+++ piptools/_compat/pip_compat.py
@@ -4,6 +4,7 @@ import optparse
 from typing import Callable, Iterable, Iterator, cast
 
 import pip
+from pip._internal.cache import WheelCache
 from pip._internal.index.package_finder import PackageFinder
 from pip._internal.network.session import PipSession
 from pip._internal.req import InstallRequirement
@@ -13,7 +14,6 @@ from pip._vendor.packaging.version import parse as parse_version
 from pip._vendor.pkg_resources import Requirement
 
 PIP_VERSION = tuple(map(int, parse_version(pip.__version__).base_version.split(".")))
-
 
 __all__ = [
     "dist_requires",
@@ -74,3 +74,10 @@ else:
             for req in reqs
             if not req.marker or req.marker.evaluate({"extra": None})
         ]
+
+
+def create_wheel_cache(cache_dir: str, format_control: str | None = None) -> WheelCache:
+    kwargs: dict[str, str | None] = {"cache_dir": cache_dir}
+    if PIP_VERSION[:2] <= (23, 0):
+        kwargs["format_control"] = format_control
+    return WheelCache(**kwargs)