+++ piptools/utils.py
@@ -7,12 +7,12 @@ import json
 import os
 import re
 import shlex
-from typing import Any, Callable, Iterable, Iterator, TypeVar, cast
+from typing import TYPE_CHECKING, Any, Callable, Iterable, Iterator, TypeVar, cast
 
 import click
 from click.utils import LazyFile
 from pip._internal.req import InstallRequirement
-from pip._internal.req.constructors import install_req_from_line
+from pip._internal.req.constructors import install_req_from_line, parse_req_from_line
 from pip._internal.utils.misc import redact_auth_from_url
 from pip._internal.vcs import is_url
 from pip._vendor.packaging.markers import Marker
@@ -21,7 +21,14 @@ from pip._vendor.packaging.utils import canonicalize_name
 from pip._vendor.packaging.version import Version
 from pip._vendor.pkg_resources import Distribution, Requirement, get_distribution
 
+from piptools._compat import PIP_VERSION
 from piptools.subprocess_utils import run_python_snippet
+
+if TYPE_CHECKING:
+    from typing import Protocol
+else:
+    Protocol = object
+
 
 _KT = TypeVar("_KT")
 _VT = TypeVar("_VT")
@@ -456,7 +463,6 @@ def copy_install_requirement(
         "markers": template.markers,
         "use_pep517": template.use_pep517,
         "isolated": template.isolated,
-        "install_options": template.install_options,
         "global_options": template.global_options,
         "hash_options": template.hash_options,
         "constraint": template.constraint,
@@ -464,6 +470,9 @@ def copy_install_requirement(
         "user_supplied": template.user_supplied,
     }
     kwargs.update(extra_kwargs)
+
+    if PIP_VERSION[:2] <= (23, 0):
+        kwargs["install_options"] = template.install_options
 
     # Original link does not belong to install requirements constructor,
     # pop it now to update later.
@@ -482,3 +491,34 @@ def copy_install_requirement(
     )
 
     return ireq
+
+
+class PackageMetadata(Protocol):
+    def get_all(self, name: str, failobj: _T = ...) -> list[str] | _T:
+        ...
+
+
+def parse_requirements_from_wheel_metadata(
+    metadata: PackageMetadata, src_file: str
+) -> Iterator[InstallRequirement]:
+    package_name = metadata.get_all("Name")[0]
+    comes_from = f"{package_name} ({src_file})"
+
+    for req in metadata.get_all("Requires-Dist") or []:
+        parts = parse_req_from_line(req, comes_from)
+        if parts.requirement.name == package_name:
+            package_dir = os.path.dirname(os.path.abspath(src_file))
+            # Replace package name with package directory in the requirement
+            # string so that pip can find the package as self-referential.
+            # Note the string can contain extras, so we need to replace only
+            # the package name, not the whole string.
+            replaced_package_name = req.replace(package_name, package_dir, 1)
+            parts = parse_req_from_line(replaced_package_name, comes_from)
+
+        yield InstallRequirement(
+            parts.requirement,
+            comes_from,
+            link=parts.link,
+            markers=parts.markers,
+            extras=parts.extras,
+        )